backend/requirements.txt
Django==5.1.2
mysqlclient==2.2.5
Pillow==10.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0